projects/3dml/requirements.txt
numpy
Pillow
scipy
ipywidgets>=8.0
ipycanvas
jupyterlab
pytest
